refactor: pipelines components 통합 및 구조 개선
